move service example to http/grpc
