+++ implement/grpc/examples/config/compare.yaml
@@ -1,0 +1,37 @@
+name: basic grpc comparison test
+destinations:
+  actual: http://localhost:50051
+  expect: http://localhost:50051
+
+testcases:
+  - target: greeter.Greeter/SayHello
+    description: wip request
+    setting:
+      allow: true
+      template:
+        name:
+          test-ap1: server1
+          test-ap2: server2
+      request:
+        message: 
+          json:
+            name: ${name}
+      response:
+        message:
+          json:
+  - target: counter.Counter/Increment
+    setting:
+      request:
+        descriptor: 
+          protos: [./dev/server/grpc/proto/counter.proto]
+          import_path: [./dev/server/grpc/proto]
+        message:
+          json:
+            100
+      response:
+        message:
+          json:
+  - target: echo.Echo/EchoMetadata
+    setting:
+      request:
+      response:
