+++ implement/http/examples/service.rs
@@ -3,23 +3,23 @@ use std::process::ExitCode;
 #[tokio::main]
 #[cfg(all(feature = "yaml", feature = "json", feature = "console-report"))]
 async fn main() -> Result<ExitCode, Box<dyn std::error::Error + Send + Sync>> {
-    use axum::{body::Body, extract::Request};
-    use relentless::{
-        implement::service_http::{evaluate::HttpResponse, factory::HttpRequest},
-        interface::command::Relentless,
-    };
+    use axum::body::Body;
+    use relentless::interface::command::{Assault, Relentless};
     use relentless_dev_server_http::route;
+    use relentless_http::command::HttpAssault;
 
-    let cmd = Relentless {
+    let assault = HttpAssault::<Body, Body>::new(Relentless {
         file: vec!["examples/config/assault.yaml".into(), "examples/config/compare.yaml".into()],
         ..Default::default()
-    };
-    let (configs, _) = cmd.configs();
-    let service = route::app_with(Default::default());
-    let report = cmd.assault_with::<HttpRequest, HttpResponse, _, Request<Body>>(configs, service).await?;
+    });
+    let (configs, errors) = assault.configs();
+    errors.into_iter().for_each(|err| eprintln!("{}", err));
 
-    cmd.report(&report)?;
-    Ok(report.exit_code(&cmd))
+    let service = route::app_with(Default::default());
+    let report = assault.assault_with(configs, service).await?;
+
+    assault.report_with(&report, std::io::stdout())?;
+    Ok(assault.exit_code(&report))
 }
 
 #[cfg(not(all(feature = "yaml", feature = "json", feature = "console-report")))]
